remove else branch from SIC DoPts
this is handled better by loading from a checkpoint now and as the old
comment noted, this was very unreliable since it required that all of
the points files were present. This would be unlikely to be the case
if run with pbqff and if not run with pbqff you can still use the
checkpoint functionality once I write the script for it. alternatively
you can now use the -freqs flag to explicitly resume from the anpass
phase after generating rel.dat