delete failing tests
This sounds bad, but these tests are better handled by full_test
without such painful and brittle setup. With the map implementation of
the symmetry, Derivatives in particular were going to rely on calling
Normalize to try to figure out the Srcs. On top of all that, they were
in the wrong file anyway