track down very subtle CartProd bug
there was some kind of issue with x getting captured in the outer loop. the
first element of tmp was changing between the start and end of the inner loop,
and this only popped up for more than 3 elements in the parts. I think for less
than that it must have been reallocating anyway